requirements.txt
dotenv~=0.9.9
python-dotenv~=1.1.1
django-googlesearch~=0.2
pandas~=2.3.0
numpy~=2.2.6
pydantic~=2.11.7
scikit-learn~=1.7.0
requests~=2.32.4
pydevd~=3.3.0
beautifulsoup4~=4.13.4
Telethon~=1.40.0
chardet~=5.2.0
googlesearch-python~=1.3.0
urllib3~=2.5.0
selenium~=4.34.2
tqdm~=4.67.1
stem~=1.8.2
tenacity~=8.5.0
openpyxl~=3.1.5
tavily-python~=0.7.11
playwright~=1.55.0